+++ lib/nylas/errors.rb
@@ -3,109 +3,83 @@
 module Nylas
   Error = Class.new(::StandardError)
 
-  class ModelActionError < Error; end
-  class ModelNotFilterableError < ModelActionError; end
-  class ModelNotCreatableError < ModelActionError; end
-  class ModelNotShowableError < ModelActionError; end
-  class ModelNotAvailableAsRawError < ModelActionError; end
-  class ModelNotListableError < ModelActionError; end
-  class ModelNotFilterableError < ModelActionError; end
-  class ModelNotSearchableError < ModelActionError; end
-  class ModelNotUpdatableError < ModelActionError; end
-  class ModelNotDestroyableError < ModelActionError; end
+  # Base error class for API-related errors.
+  class AbstractNylasApiError < Error; end
 
-  class JsonParseError < Error; end
+  # Base error class for SDK-related errors.
+  class AbstractNylasSdkError < Error; end
 
-  # Raised when attempting to set a field that is not on a model with mass assignment
-  class ModelMissingFieldError < ModelActionError
-    def initialize(field, model)
-      super("#{field} is not a valid attribute for #{model.class.name}")
-    end
-  end
+  # Error class representing a failed parse of a JSON response from the Nylas API.
+  class JsonParseError < AbstractNylasSdkError; end
 
-  # Indicates that a given method needs an access token to work.
-  class NoAuthToken < Error
-    def initialize(method_name)
-      super "No access token was provided and the #{method_name} method requires one"
-    end
-  end
+  # Error class representing a failed response from the Nylas API.
+  class NylasApiError < AbstractNylasApiError
+    attr_accessor :type, :request_id, :provider_error, :status_code
 
-  UnexpectedAccountAction = Class.new(Error)
-  UnexpectedResponse = Class.new(Error)
-
-  # Base class to inflate the standard errors returned from the Nylas API
-  class APIError < Error
-    attr_accessor :type
-    attr_accessor :message
-    attr_accessor :server_error
-
-    def initialize(type, message, server_error = nil)
+    # Initializes an error and assigns the given attributes to it.
+    #
+    # @param type [Hash] Error type.
+    # @param message [String] Error message.
+    # @param status_code [Integer] Error status code.
+    # @param provider_error [String, nil] Provider error.
+    # @param request_id [Hash, nil] The ID of the request.
+    def initialize(type, message, status_code, provider_error = nil, request_id = nil)
       super(message)
       self.type = type
-      self.message = message
-      self.server_error = server_error
+      self.status_code = status_code
+      self.provider_error = provider_error
+      self.request_id = request_id
     end
 
-    def self.parse_error_response(response)
+    # Parses the error response.
+    #
+    # @param response [Hash] Response from the Nylas API.
+    # @param status_code [Integer] Error status code.
+    def self.parse_error_response(response, status_code)
       new(
         response["type"],
         response["message"],
-        response["server_error"]
+        status_code,
+        response["provider_error"]
       )
     end
   end
 
-  # Error class representing a 429 error response, with details on the rate limit
-  class RateLimitError < APIError
-    attr_accessor :rate_limit
-    attr_accessor :rate_limit_reset
+  # Error class representing a failed response from the Nylas OAuth integration.
+  class NylasOAuthError < AbstractNylasApiError
+    attr_accessor :error, :error_description, :error_uri, :error_code, :status_code
 
-    RATE_LIMIT_LIMIT_HEADER = "x_ratelimit_limit"
-    RATE_LIMIT_RESET_HEADER = "x_ratelimit_reset"
-
-    def initialize(type, message, server_error = nil, rate_limit = nil, rate_limit_reset = nil)
-      super(type, message, server_error)
-      self.rate_limit = rate_limit
-      self.rate_limit_reset = rate_limit_reset
+    # Initializes an error and assigns the given attributes to it.
+    #
+    # @param error [String] Error type.
+    # @param error_description [String] Description of the error.
+    # @param error_uri [String] Error URI.
+    # @param error_code [String] Error code.
+    # @param status_code [String] Error status code.
+    def initialize(error, error_description, error_uri, error_code, status_code)
+      super(error_description)
+      self.error = error
+      self.error_description = error_description
+      self.error_uri = error_uri
+      self.error_code = error_code
+      self.status_code = status_code
     end
-
-    def self.parse_error_response(response)
-      rate_limit, rate_limit_rest = extract_rate_limit_details(response)
-
-      new(
-        response["type"],
-        response["message"],
-        response["server_error"],
-        rate_limit,
-        rate_limit_rest
-      )
-    end
-
-    def self.extract_rate_limit_details(response)
-      return nil, nil unless response.respond_to?(:headers)
-
-      rate_limit = response.headers[RATE_LIMIT_LIMIT_HEADER.to_sym].to_i
-      rate_limit_rest = response.headers[RATE_LIMIT_RESET_HEADER.to_sym].to_i
-
-      [rate_limit, rate_limit_rest]
-    end
-
-    private_class_method :extract_rate_limit_details
   end
 
-  AccessDenied = Class.new(APIError)
-  ResourceNotFound = Class.new(APIError)
-  MethodNotAllowed = Class.new(APIError)
-  InvalidRequest = Class.new(APIError)
-  UnauthorizedRequest = Class.new(APIError)
-  ResourceRemoved = Class.new(APIError)
-  TeapotError = Class.new(APIError)
-  RequestTimedOut = Class.new(APIError)
-  MessageRejected = Class.new(APIError)
-  SendingQuotaExceeded = Class.new(RateLimitError)
-  ServiceUnavailable = Class.new(APIError)
-  BadGateway = Class.new(APIError)
-  InternalError = Class.new(APIError)
-  EndpointNotYetImplemented = Class.new(APIError)
-  MailProviderError = Class.new(APIError)
+  # Error class representing a timeout from the Nylas SDK.
+  class NylasSdkTimeoutError < AbstractNylasSdkError
+    attr_accessor :url, :timeout
+
+    # Initializes an error and assigns the given attributes to it.
+    # @param url [String] URL that timed out.
+    # @param timeout [Integer] Timeout in seconds.
+    # @return [NylasSdkTimeoutError] The error object.
+    def initialize(url, timeout)
+      super("Nylas SDK timed out before receiving a response from the server.")
+      self.url = url
+      self.timeout = timeout
+    end
+  end
+
+  HTTP_SUCCESS_CODES = [200, 201, 202, 302].freeze
 end