+++ spec/nylas/free_busy_collection_spec.rb
@@ -4,6 +4,7 @@ require "spec_helper"
 
 describe Nylas::FreeBusyCollection do
   describe "#each" do
+    # Get a collection of FreeBusy objects.
     it "returns collection of `FreeBusy` objects" do
       start_time = 1_609_439_400
       end_time = 1_640_975_400